+++ src/Inventory.php
@@ -37,11 +37,6 @@ class Inventory extends Plugin
     public static Plugin $plugin;
 
     /**
-     * @var bool The plugin has its own section.
-     */
-    public bool $hasCpSection = true;
-
-    /**
      * @inheritdoc
      */
     public function init(): void
@@ -55,7 +50,7 @@ class Inventory extends Plugin
         ]);
 
         // Load class services
-        $inventoryService = Inventory::$plugin->inventoryService;
+        $inventoryService = self::$plugin->inventoryService;
 
         // Set template hook
         Craft::$app->getView()->hook('getFieldLayouts', [$inventoryService, 'getFieldLayouts']);
@@ -66,31 +61,6 @@ class Inventory extends Plugin
 
         // Redirect after plugin install
         $this->_postInstallRedirect();
-
-        // Manage whether "Inventory" should appear in the sidebar
-        $this->_manageSidebarNav();
-    }
-
-    // ========================================================================= //
-
-    /**
-     * Manage whether "Inventory" should appear in the sidebar.
-     */
-    private function _manageSidebarNav(): void
-    {
-        // If not a web request, bail
-        if (!Craft::$app->getRequest()->getIsCpRequest()) {
-            return;
-        }
-
-        // Whether a cookie has been set to hide "Inventory" from the sidebar
-        $hideFromSidebar = Craft::$app->getRequest()->getCookies()->getValue('hide-inventory-from-sidebar');
-
-        // If cookie exists
-        if ($hideFromSidebar) {
-            // Hide "Inventory" from the sidebar
-            $this->hasCpSection = false;
-        }
     }
 
     // ========================================================================= //
@@ -129,7 +99,7 @@ class Inventory extends Plugin
         Event::on(
             UrlManager::class,
             UrlManager::EVENT_REGISTER_CP_URL_RULES,
-            function (RegisterUrlRulesEvent $event) {
+            static function (RegisterUrlRulesEvent $event) {
                 $event->rules['utilities/inventory/<groupId:\d+>'] = ['template' => 'inventory/fields'];
             }
         );